added tables to DB: GroupPages, Members and Socials

+++ StoronnimV/StoronnimV.Contracts/Repositories/IRepository.cs
@@ -5,7 +5,7 @@ namespace StoronnimV.Contracts.Repositories;
 
 public interface IRepository<T> where T : BaseEntity
 {
-    Task<T?> GetByIdAsync(BigInteger id);
+    Task<T?> GetByIdAsync(long id);
     Task<IEnumerable<T>?> GetAllAsync();
     Task AddAsync(T entity);
     Task UpdateAsync(T entity, Action updateAction);
